move ClientInputs to model folder
